+++ react/typings/vtex.session-client.d.ts
@@ -1,0 +1,2 @@
+declare module 'vtex.session-client'
+declare module 'vtex.session-client/*'