+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support show for multiple shipping costs by commercial policy in `react/components/MinicartFreeshipping.tsx` file
+
 ## [1.2.0] - 2022-03-21
 
 ### Added